Options/optiontickers path MaunaloaController

+++ src/main/java/harborview/maunaloa/MaunaloaModel.java
@@ -100,7 +100,7 @@ public class MaunaloaModel {
     }
     private Chart volumeChart(List<StockPrice> spots) {
         Chart chart = new Chart();
-        List<Integer> vol = spots.stream().map(x -> x.getVolume()).collect(Collectors.toList());
+        List<Double> vol = spots.stream().map(x -> (double)x.getVolume()).collect(Collectors.toList());
         OptionalDouble maxVol = vol.stream().mapToDouble(v -> v).max();
         maxVol.ifPresent(v -> {
            List<Double> normalized = vol.stream().map(x -> x/v).collect(Collectors.toList());
